Further explain "all mailboxen" problem
Some kind of tree graph would help in understanding messaging patterns,
ie, which message types go to which mailboxen.  A simple solution is to
use `bus.subscribersTo<Nothing>()` as `Nothing` is a subtype of all
types, however, `Nothing` is not instantiable nor reifiable as a
`Class`.

Current workaround is to use `bus.subscribers` property, and manually
inspect the FIFO-ordered map.